HashEntry: allow __eq__ against os.stat_result
I was about to add a type comparison to the relevant line in
HashDatabase.update, but this started to get a bit unwieldy. Rather than adding
all of the logic to this line, allow direct equality comparisons between
HashEntry objects and stat_result objects.

If you pass something that isn't a stat_result to HashEntry.__eq__, it falls
back to the normal object.__eq__ behavior.